+++ Flask_Marketplace/__init__.py
@@ -11,6 +11,6 @@ __version__ = '1.0.0-alpha'
 
 import os, sys
 sys.path.append(os.path.dirname(__file__))
-from Flask_Marketplace.shop_module.views_class import MarketViews
+from Flask_Marketplace.views import MarketViews
 from .factory import marketplace
