+++ Flask_Marketplace/forms/user_forms.py
@@ -3,7 +3,6 @@ from flask_wtf import FlaskForm
 from wtforms import SubmitField, StringField, ValidationError
 from wtforms.validators import required
 
-from Flask_Marketplace.users_module.models import User
 
 
 def unique_entry(form, field):
