+++ Flask_Marketplace/models/__init__.py
@@ -1,5 +1,5 @@
 ''' Module for handling interactions '''
 from flask_security import SQLAlchemyUserDatastore
-from Flask_Marketplace.users_module.models import db, User, Role, UserAdmin, RoleAdmin
+from Flask_Marketplace.models.user_models import db, User, Role, UserAdmin, RoleAdmin
 
 user_datastore = SQLAlchemyUserDatastore(db, User, Role)
