--- Flask_Marketplace/shop_module/__init__.py
@@ -1,1 +1,0 @@
-''' Module for handling shop products interactions '''
