+++ plugins/flw_module/__init__.py
@@ -1,4 +1,2 @@
 from .models import FlwDetail
 from .forms import AccountForm
-
-from .forms import MarketViews
