+++ plugins/flw_module/forms.py
@@ -4,7 +4,7 @@ from wtforms import (IntegerField, SelectField, StringField,
 from wtforms.validators import required
 
 from .utilities import bank_options
-from Flask_Marketplace.shop_module.forms import AccountForm
+from Flask_Marketplace.forms.shop_forms import AccountForm
 
 
 class AccountForm(AccountForm):
@@ -15,12 +15,3 @@ class AccountForm(AccountForm):
                                   'data-live-search': 'true'}
                        )
     save = SubmitField()
-
-from Flask_Marketplace import MarketViews as ujik
-from flask import render_template
-class MarketViews(ujik):
-    def index(self):
-        print("asdfghjklpuy What")
-        latest = self.Product.query.limit(3).all()
-        return render_template('marketplace/h8ome.html', products=latest)
-
